Configure the project for continuous integration (CI)
Add the .travis.yml to the root directory of the project
to configure the project to work with the Travis CI build
automation. This will allow the project ot be build and tested
everytime the is a code change to ensure that the changes
do not break the working code.